requirements.txt
streamlit
google-generativeai
requests
beautifulsoup4
pandas
numpy
newspaper3k
lxml
plotly
openpyxl
lxml[html_clean]
python-Levenshtein
fuzzywuzzy